+++ test/sol4/kyberProxyV1.js
@@ -5,7 +5,7 @@ const MatchingEngine = artifacts.require("KyberMatchingEngine.sol");
 const KyberStorage = artifacts.require("KyberStorage.sol");
 const RateHelper = artifacts.require("KyberRateHelper.sol");
 const TestToken = artifacts.require("TestToken.sol");
-const MaliciousReserve = artifacts.require("MaliciousReserve.sol");
+const MaliciousReserve = artifacts.require("MaliciousReserve2.sol");
 const KyberNetwork = artifacts.require("./KyberNetwork.sol");
 const NetworkNoMaxDest = artifacts.require("KyberNetworkNoMaxDest.sol");
 const MaliciousNetwork = artifacts.require("MaliciousKyberNetwork.sol");
