+++ test/sol6/fuzzerFiles/stakingFuzzer/stakingActionsGenerator.js
