+++ test/sol6/fuzzerFiles/stakingFuzzer/stakingSimulator.js
