+++ test/sol6/fuzzerFiles/tradeFuzzer/multiple_trade_fuzz_test.sh
