+++ test/sol6/fuzzerFiles/tradeFuzzer/networkSimulator.js
@@ -8,8 +8,8 @@ const { expectRevert } = require('@openzeppelin/test-helpers');
 
 const winston = require('winston');
 
-const TradeParamGenerator = require("../trades/tradeParamsGenerator.js");
-const { TRADE, UPDATE_RESERVE_RATE, RevertType, allRevertMessages } = require("../trades/tradeParamsGenerator.js");
+const TradeParamGenerator = require("./tradeParamsGenerator.js");
+const { TRADE, UPDATE_RESERVE_RATE, RevertType, allRevertMessages } = require("./tradeParamsGenerator.js");
 
 let numberSuccessfulTrades = 0;
 let numberGettingsZeroRates = 0;
